# docs/openapi.yaml
openapi: 3.1.0
info:
  title: AgentCordon API
  version: 1.6.1
  description: |
    AgentCordon is a self-hostable Agentic Identity Provider (IdP) that provides
    secure, auditable, policy-controlled token brokerage for autonomous agents.
    Agents do not directly hold long-lived credentials; every issuance and use
    is observable via the audit log.
  contact:
    name: AgentCordon
  license:
    name: MIT
servers:
  - url: http://localhost:3140
    description: Local development server
...
security:
  - SessionCookie: []
  - BearerJWT: []
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: agtcrdn_session
      description: |
        Session cookie set after user login. Every unsafe method authenticated
        this way is also CSRF-protected by a double-submit rule: the request
        must carry an `X-CSRF-Token` header whose value matches the readable
        `agtcrdn_csrf` cookie, or it is refused with
        `403 csrf_validation_failed`.
    BearerJWT:
      type: http
      scheme: bearer
      description: |
        An **opaque** OAuth 2.0 access token issued to a workspace by
        `POST /api/v1/oauth/token`. Despite the scheme name this is not a JWT
        and carries no claims a client can read: the server stores only its
        SHA-256 hash and resolves the workspace by token id. The ES256 JWT
        issuer, its JWKS endpoint, and the permissions-token endpoint were
        removed.
...
paths:
...
  /api/v1/auth/oidc/callback:
    get:
      tags: [OIDC]
      summary: OIDC callback
      description: |
        Handles the IdP callback after authorization. Validates the state,
        exchanges the code for tokens, validates the ID token, provisions or
        finds the user, creates a session, and redirects to the UI.
      security: []
      parameters:
        - name: code
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
        - name: error
          in: query
          schema:
            type: string
        - name: error_description
          in: query
          schema:
            type: string
      responses:
        '302':
          description: Redirect to UI with session cookies set
...
  /api/v1/mcp-servers/oauth/callback:
    get:
      tags: [MCP]
      summary: OAuth callback handler for MCP servers
      description: |
        Handles the OAuth2 callback from the identity provider after user
        authorization. Exchanges the authorization code for tokens and stores
        them for the MCP server.
      security: []
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
      responses:
        '302':
          description: Redirects to MCP server detail page on success
        '400':
          description: Invalid or expired state/code
...
  /api/v1/oauth-provider-clients:
    get:
      tags: [MCP]
      summary: List OAuth provider clients
      description: |
        Requires `manage_mcp_servers`, so an operator can see which client an
        origin uses. The client secret is never part of the response.
      security:
        - SessionCookie: []
      responses:
        '200': { description: Provider clients (client secrets never returned) }
        '403': { description: Caller does not hold manage_mcp_servers }
    post:
      tags: [MCP]
      summary: Create an OAuth provider client by hand
      description: |
        For an authorization server that does not support RFC 7591 Dynamic
        Client Registration. `authorization_server_url` is the lookup key and
        must be the bare origin (`scheme://host[:port]`, lowercase, no path,
        no trailing slash) — discovery normalizes the value it reads from the
        resource metadata and compares it to this field as stored.

        Requires `manage_oauth_provider_clients`, which the default policy
        grants to enabled admins only — the row is shared by every tenant with
        an MCP server at that origin.
      security:
        - SessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [authorization_server_url, client_id]
              properties:
                label: { type: string }
                authorization_server_url: { type: string, format: uri }
                authorize_endpoint: { type: string, format: uri }
                token_endpoint: { type: string, format: uri }
                client_id: { type: string }
                client_secret: { type: string }
                scopes: { type: string }
      responses:
        '201': { description: Provider client created }
        '403': { description: Caller does not hold manage_oauth_provider_clients }
        '409': { description: A client already exists for that authorization server origin }
  /api/v1/oauth-provider-clients/{id}:
    get:
      tags: [MCP]
      summary: Get an OAuth provider client
      description: Requires `manage_mcp_servers`. The client secret is never returned.
      security:
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Provider client }
        '403': { description: Caller does not hold manage_mcp_servers }
        '404': { description: Not found }
    put:
      tags: [MCP]
      summary: Update an OAuth provider client
      description: |
        Requires `manage_oauth_provider_clients` (admins only). Allowed even
        while credentials and MCP servers depend on the row — this is how a
        rotated provider secret gets in. The audit event names which fields
        changed and how many credentials and MCP servers were depending on the
        client, never the secret value.
      security:
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Updated }
        '403': { description: Caller does not hold manage_oauth_provider_clients }
        '404': { description: Not found }
        '409': { description: The row was created by dynamic client registration; re-register it instead }
    delete:
      tags: [MCP]
      summary: Delete an OAuth provider client
      description: |
        Requires `manage_oauth_provider_clients` (admins only). Refused while
        anything still authenticates at that origin.
      security:
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Deleted }
        '403': { description: Caller does not hold manage_oauth_provider_clients }
        '404': { description: Not found }
        '409':
          description: |
            Still in use. The message names both counts — the
            `oauth2_user_authorization` credentials issued against this
            authorization server, and the OAuth2 MCP servers that authenticate
            with them.
  /api/v1/oauth-provider-clients/{id}/reregister:
    post:
      tags: [MCP]
      summary: Re-run discovery and dynamic client registration
      description: Requires `manage_oauth_provider_clients` (admins only).
      security:
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Row updated in place }
        '403': { description: Caller does not hold manage_oauth_provider_clients }
        '404': { description: Not found }

  # ============================================================================
  # OAuth 2.0 authorization server
  # ============================================================================
...